ROM-84 Containerコンポーネントを実装する。
