+++ src/components/layouts/Container/styles.ts
@@ -1,0 +1,10 @@
+import { css } from '@emotion/react';
+
+import { MAX_WINDOW_WIDTH, MIN_WINDOW_WIDTH } from '@/app/theme/breakPoints';
+
+export const styles = css({
+  backgroundColor: 'inherit',
+  width: 'inherit',
+  minWidth: MIN_WINDOW_WIDTH,
+  maxWidth: MAX_WINDOW_WIDTH,
+});
