+++ src/components/layouts/Header/DesktopLayout/styles.ts
@@ -7,13 +7,16 @@ import { spaces } from '@/app/theme/spaces';
 export const styles = {
   container: css({
     display: 'flex',
+    justifyContent: 'center',
+    backgroundColor: colors.white,
+    borderBottom: `solid ${colors.blackHaze} ${borderWidths.m}px`,
+  }),
+  innerContainer: css({
+    display: 'flex',
     justifyContent: 'space-between',
     alignItems: 'center',
-    width: '100%',
     height: 60,
-    backgroundColor: colors.white,
     padding: `0 ${spaces.m}px`,
-    borderBottom: `solid ${colors.blackHaze} ${borderWidths.m}px`,
   }),
   navWrapper: css({
     display: 'flex',
