+++ src/components/layouts/Header/MobileLayout/styles.ts
@@ -9,13 +9,17 @@ import { spaces } from '@/app/theme/spaces';
 export const styles = {
   container: css({
     display: 'flex',
+    justifyContent: 'center',
+    backgroundColor: colors.white,
+    borderBottom: `solid ${colors.blackHaze} ${borderWidths.m}px`,
+  }),
+  innerContainer: css({
+    display: 'flex',
     justifyContent: 'space-between',
     alignItems: 'center',
     width: '100%',
     height: 50,
-    backgroundColor: colors.white,
     padding: `0 ${spaces.s}px`,
-    borderBottom: `solid ${colors.blackHaze} ${borderWidths.m}px`,
   }),
 
   detailBtn: css({
